+++ src/main/java/com/bttf/queosk/dto/MenuCreationForm.java
@@ -1,4 +1,4 @@
-package com.bttf.queosk.dto.menudto;
+package com.bttf.queosk.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
