+++ src/main/java/com/bttf/queosk/dto/MenuStatusForm.java
@@ -1,5 +1,6 @@
-package com.bttf.queosk.dto.menudto;
+package com.bttf.queosk.dto;
 
+import com.bttf.queosk.enumerate.MenuStatus;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
@@ -9,7 +10,6 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 @Builder
-public class MenuUpdateForm {
-    private String name;
-    private Long price;
+public class MenuStatusForm {
+    private MenuStatus status;
 }
