+++ src/main/java/com/bttf/queosk/dto/MenuUpdateForm.java
@@ -1,6 +1,5 @@
-package com.bttf.queosk.dto.menudto;
+package com.bttf.queosk.dto;
 
-import com.bttf.queosk.enumerate.MenuStatus;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
@@ -10,11 +9,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 @Builder
-public class MenuDto {
-    private Long id;
-    private Long restaurantId;
+public class MenuUpdateForm {
     private String name;
-    private String imageUrl;
     private Long price;
-    private MenuStatus status;
 }
