+++ src/main/java/com/bttf/queosk/entity/Menu.java
@@ -1,5 +1,6 @@
 package com.bttf.queosk.entity;
 
+import com.bttf.queosk.dto.MenuCreationForm;
 import com.bttf.queosk.entity.baseentity.BaseTimeEntity;
 import com.bttf.queosk.enumerate.MenuStatus;
 import lombok.AllArgsConstructor;
@@ -9,6 +10,8 @@ import lombok.NoArgsConstructor;
 import org.hibernate.envers.AuditOverride;
 
 import javax.persistence.*;
+
+import static com.bttf.queosk.enumerate.MenuStatus.ON_SALE;
 
 @Entity(name = "menu")
 @AuditOverride(forClass = BaseTimeEntity.class)
@@ -46,5 +49,15 @@ public class Menu extends BaseTimeEntity {
 
     public void setStatus(MenuStatus status) {
         this.status = status;
+    }
+
+    public static Menu of(Long restaurantId, MenuCreationForm menuCreationForm) {
+        return Menu.builder()
+                .name(menuCreationForm.getName())
+                .price(menuCreationForm.getPrice())
+                .status(ON_SALE)
+                .imageUrl(menuCreationForm.getImageUrl())
+                .restaurantId(restaurantId)
+                .build();
     }
 }
